+++ src/vertex.cpp
@@ -1,6 +1,6 @@
-#include "vertex.h"
+#include "headers/vertex.h"
 
-Vertex::Vertex(float x, float y, float z, float r, float g, float b)
+Vertex::Vertex(double x, double y, double z, double r, double g, double b)
     : m_x {x}
     , m_y {y}
     , m_z {z}
@@ -8,7 +8,7 @@ Vertex::Vertex(float x, float y, float z, float r, float g, float b)
 {
 }
 
-Vertex::Vertex(float x, float y, float z, Color c)
+Vertex::Vertex(double x, double y, double z, Color c)
     : m_x {x}
     , m_y {y}
     , m_z {z}
